+++ api/src/numerical-guidance/application/query/indicator/basic/dto/indicators.dto.ts
