+++ api/src/numerical-guidance/application/query/indicator/basic/get-indicator/get-indicators.query.ts
