+++ api/src/numerical-guidance/infrastructure/adapter/persistence/history-indicator/mapper/history-indicator.mapper.ts
@@ -1,8 +1,8 @@
 import { HistoryIndicatorValueEntity } from '../../history-indicator-value/entity/history-indicator-value.entity';
 import { HistoryIndicatorDto } from '../../../../../application/query/history-indicator/dto/history-indicator.dto';
-import { Indicator } from '../../../../../application/query/indicator/dto/indicator.dto';
+import { Indicator } from '../../../../../application/query/indicator/basic/dto/indicator.dto';
 import { HistoryIndicatorEntity } from '../entity/history-indicator.entity';
-import { IndicatorValue } from '../../../../../application/query/live-indicator/dto/live-indicator.dto';
+import { IndicatorValue } from '../../../../../../utils/type/type-definition';
 
 export class HistoryIndicatorMapper {
   static mapEntitiesToDto(historyIndicatorEntity: HistoryIndicatorEntity, values: IndicatorValue[]) {
